requirements.txt
# requirements.txt
# Environment: Python 3.11

# Data Manipulation
pandas>=2.0.0
numpy>=1.24.0,<2.0.0

# Configuration
PyYAML>=6.0

# Machine Learning & Metrics
scikit-learn>=1.2.0
scipy>=1.10.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Testing
pytest>=7.0.0